Starting work on representing this in scala
It feels like we can get to a point where the representation is easy to
edit even for non-programmers, with the big benefit that this structure
provides:
1. Automatic checking/verifying that the configuration is valid.
2. An easy way to parse the structure for a tool that converts it into
whatever representation we like.
3. There is some precedent in the RISC-V community of using scala in the
form of chisel. I know chisel isn't popular, but at least some hardware
designers will have some familiarity with it already.