Add custom support for fixed-width numbers. (#24)
* Refactor code. Doesn't change size.

* Add support for fixed-length numbers.

Use it for maskmax, which increases the example size by a byte because
we're putting 0 in there. (0 encoded as a variable length number is
fewer bits than encoded in an Int6.)

Fixed numbers are not yet supported by the C decoder.

* Implement fixed-length decoding in C.

Somehow the code got a lot smaller. I don't understand why.

New size: 1819/1873