Get rid of all non-ASN.1 stuff.
Implement a simple converter in asn1tools that can be used to convert
between the examples files and UPER.

Update the examples so they are suitable as input to asn1tools. This
means that lists have to have [], even if they only have a single entry,
and every value must be mentioned. I'd like to remove those restrictions
in the future.