Use SEQUENCE of CHOICE types instead of OPTIONAL. (#28)
* Add `test` subcommand to rvcs.py

This converts the given files to UPER, converts them back, and finally
checks that the input matches the final output.

Example:
```
tnewsome@compy-linux:~/SiFive/configuration-structure$ ./asn1tools/rvcs.py test examples/*.yaml
examples/debug_module.yaml is 20B in UPER
{'dictionary_item_removed': {"root['harts'][4]['fastInt']",
                             "root['traceModule']['branchPredictorEntries']",
                             "root['traceModule']['contextBusWidth']",
                             "root['traceModule']['jumpTargetCacheEntries']"}}
```

As you can see our example YAML doesn't pass the test, because it
doesn't match the capitalization in the schema.

* Fix capitalization.

```
tnewsome@compy-linux:~/SiFive/configuration-structure$ ./asn1tools/rvcs.py test examples/*.yaml
examples/debug_module.yaml is 20B in UPER
examples/example.yaml is 110B in UPER
```

* Use SEQUENCE of CHOICE types instead of OPTIONAL.

We're concerned that we'll end up with a ton of hart extensions, and
each optional one will use a bit even if not mentioned at all. A
sequence of choices scales better.

With our current example, the binary size increases from 110B to 116B.
However, if I add 100 dummy extensions, the OPTIONAL version becomes
172B while the SEQUENCE of CHOICE version only increases to 119B. I
think it is worth getting this scalability. We might want to use the
same structure in a few other places as well.

There is another problem. asn1tools refuses to use CHOICE with trees
that come out of python's yaml loader. The problem is that it wants a
CHOICE to be represented as a tuple, and the yaml loader turns them into
a list. To work around this I converted the yaml to jer. jer is more
constrained and doesn't allow comments, but for now at least it works.
Long term we should figure out a way to parse something friendlier.

examples/debug_module.jer is 20B in UPER
examples/example.jer is 116B in UPER

Note: This commit builds on #27, so that should be reviewed/merged
first.