Add a tiny header for future proofing.
ASN.1 lets us add fields anywhere we marked things extensible, but what
if in some future we want to switch to a different ASN.1 encoding, or
even off of ASN.1 altogether?

This change adds a 2-bit header that will expand to be minimal overhead
if we want to switch to something else in the future (9 bits, I think).
It's unlikely that we will go that route, but this prepares us just in
case.

Also see mailing list subject:"configuration structure header"