Use SEQUENCE OF CHOICE instead of OPTIONAL.
This reduces the size of small descriptions (of presumably small
systems), but increases the size of larger descriptions.

Assuming we're going to end up with 53 or so extensions, descriptions
that use 8 or fewer of them will benefit up to 6 bytes from this change.
Descriptions that use 26 extensions in this example would pay 19 bytes.
(If all extensions are used, the cost of this change is 45 bytes.)

Is that a worthwhile trade-off?

In our specific examples, the numbers change:
examples/debug_module.jer stays the same at 18B
examples/example.jer grows from 93B to 97B