Make rvcs C library using asn1c (#105)
* WIP on CS parser

Builds/links for RV64. It doesn't for RV32, but I think that's because I
don't have the right libc etc. installed on my system to link against.

There's nothing useful in example.c yet.

* Build library in github workflow.

* Install meson.

* Install cross build packages.

* apt get update first

* That's `apt update`

* Compile for unknown, not Linux

* Don't use new flag in asn1c.

* Add native build.

* Work with asn1c 0.9.28.

That's the one in the Ubuntu package.

* Try building a Linux binary.

In theory the Ubuntu packages include enough to do that, even if they
don't include enough for riscv64-unknown-elf binaries.

* It's riscv64-linux-gnu, not riscv64-linux-elf

* Complete simple example.

This binary reads a UPER configuration structure, and prints it out in
ASN format.

* Create RISC-V abstraction layer.

This works, but the C structure you get back is not convenient to use at
all.

* Create a tar file to distribute the library.

This should include everything a user of the C library needs.

It also contains a bunch of unnecessary C files, unfortunately.

* Try to install meson using pip3.

* Now need to manually install ninja.

* Upload library after build.

* Don't be verbose when making archive.

* Figuring out where the artifact ends up

* Add dummy in.h

The ASN code wants to include netinet/in.h, but doesn't appear to use
anything from that file?

* Add more examples/instructions to README files.